+++ input_others.go
